move src/cdp/stubTransport.spec.ts -> src/utils/transportStub.spec.ts
