+++ src/cdp/cdpClient.spec.ts
@@ -21,8 +21,9 @@ import * as sinon from 'sinon';
 import {Protocol} from 'devtools-protocol';
 import chaiAsPromised from 'chai-as-promised';
 
+import {StubTransport} from '../utils/transportStub.spec.js';
+
 import {CdpConnection} from './cdpConnection.js';
-import {StubTransport} from './stubTransport.spec.js';
 
 chai.use(chaiAsPromised);
